Move models into scheme folder
