+++ app/models/water/billing-charge-category.model.js
@@ -7,7 +7,7 @@
 
 const { Model } = require('objection')
 
-const BaseModel = require('./base.model.js')
+const BaseModel = require('../base.model.js')
 
 class BillingChargeCategoryModel extends BaseModel {
   static get tableName () {
