+++ app/models/water/licence.model.js
@@ -7,7 +7,7 @@
 
 const { Model } = require('objection')
 
-const BaseModel = require('./base.model.js')
+const BaseModel = require('../base.model.js')
 
 class LicenceModel extends BaseModel {
   static get tableName () {
